better tests structure
